// tools/conformance/src/modules/jupyter-notebook-application-generator/jupyter-notebook-application-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface JupyterNotebookApplicationGeneratorArguments {
  options: JupyterNotebookApplicationGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface JupyterNotebookApplicationGeneratorOptions {
  description?: string;
  name?: string;
}